Fix splitting names with escaped space characters
In LaTeX, "\ " is a "control space" character, which is more or less
equivalent to a single space, except that the space always inserted
in the output literally and not ignored or merged with other spaces.

BibTeX does not treat "\ " specially. The name "Christina A. L.\ Thiele"
would be split as "Christina", "A.", "L.\", "Thiele".  The stray
backslash at the end of string is not valid TeX and is causing problems
with latexcodec. So Pybtex now treats "\ " as a single space character
when splitting names.  The name "Christina A. L.\ Thiele" is now split
as "Christina", "A.", "L.", "Thiele", without the backslash.